+++ Sources/KeyboardKit/Autocomplete/AutocompleteToolbarItem.swift
@@ -23,7 +23,7 @@ public struct AutocompleteToolbarItem: View {
      
      - Parameters:
        - suggestion: The suggestion to display in the view.
-       - style: The style to apply to the item, by default `.standard`.
+       - style: The style to apply to the item, by default ``AutocompleteToolbarItemStyle/standard``.
        - locale: The locale to use to resolve quotation.
      */
     public init(
@@ -92,8 +92,8 @@ struct AutocompleteToolbarItem_Previews: PreviewProvider {
     }
     
     static let previewSuggestions: [AutocompleteSuggestion] = [
-        StandardAutocompleteSuggestion("Foo", isUnknown: true),
-        StandardAutocompleteSuggestion("Bar", isAutocomplete: true),
+        StandardAutocompleteSuggestion(text: "Foo", isUnknown: true),
+        StandardAutocompleteSuggestion(text: "Bar", isAutocomplete: true),
         StandardAutocompleteSuggestion(text: "", title: "Baz", subtitle: "Recommended")]
 }
 #endif