+++ Sources/KeyboardKit/Autocomplete/AutocompleteToolbarSeparator.swift
@@ -18,7 +18,7 @@ public struct AutocompleteToolbarSeparator: View {
      Create an autocomplete toolbar item separator.
      
      - Parameters:
-       - style: The style to apply to the separator line, by default `.standard`.
+       - style: The style to apply to the separator line, by default ``AutocompleteToolbarSeparatorStyle/standard``.
      */
     public init(style: AutocompleteToolbarSeparatorStyle = .standard) {
         self.style = style