+++ Sources/KeyboardKit/_Deprecated/Feedback+Deprecated.swift
@@ -1,5 +1,6 @@
 import Foundation
 
+
 @available(*, deprecated, renamed: "AudioFeedback")
 public typealias SystemAudio = AudioFeedback
 
@@ -30,4 +31,32 @@ public extension AudioFeedbackEngine {
 
     @available(*, deprecated, renamed: "trigger")
     func play(_ feedback: AudioFeedback) { trigger(feedback) }
+}
+
+
+
+@available(*, deprecated, renamed: "HapticFeedbackEngine")
+public typealias HapticFeedbackPlayer = HapticFeedbackEngine
+
+#if os(iOS)
+@available(*, deprecated, renamed: "StandardHapticFeedbackEngine")
+public typealias StandardHapticFeedbackPlayer = StandardHapticFeedbackEngine
+#endif
+
+@available(*, deprecated, renamed: "DisabledHapticFeedbackEngine")
+public typealias DisabledHapticFeedbackPlayer = DisabledAudioFeedbackEngine
+
+public extension HapticFeedback {
+
+    @available(*, deprecated, renamed: "engine")
+    static var player: HapticFeedbackEngine {
+        get { engine }
+        set { engine = newValue }
+    }
+}
+
+public extension HapticFeedbackEngine {
+
+    @available(*, deprecated, renamed: "trigger")
+    func play(_ feedback: HapticFeedback) { trigger(feedback) }
 }